refactored window handling
the "listwindow" and the "previewwindow" are now handled
in a more intelligent way: instead of re-creating a new
buffer and its content for each of both on each (re-)open,
now a buffer for the content of both windows is just created
one time and then only closed and re-opened in new split windows.

further from now on option "nomodifiable" is used to ensure
that the user cannot change content of "listwindow" or "previewwindow"
content.