bugfix of broken highlightning after 'R'
sometimes after refreshing the targetlist with 'R' and
when the content of the target list then really changed,
some problems ocurred with line highlightning when trying
to mark a target line with 'space'. the line was marked
but it was not highlighted.

I was not able to reproduce this reliably, hohever I guess
the list s:buildMenu.markedTargets might not be correctly
been cleaned up in those cases.

This commit attempts to fix this issue my cleaning
s:buildMenu.markedTargets up reliable when pressing 'R'.